+++ CHANGES.rst
@@ -2,8 +2,14 @@
 CHANGELOG
 =========
 
+Unreleased
+==========
+
+- The format for Redis keys for user flags was changed from
+  ``FLAG:user:<uid>:<FLAG_NAME_UPPER>`` to ``FLAG:<FLAG_NAME_UPPER>:user:<uid>``.
+
 0.1.0 - 2020-01-22
 ==================
 
 - Initial release. There are "Flags", "Samples", and "Switches" to control your
-  application. 
+  application.